refactor: extract invalid claim flow helpers
