+++ aggsender/flows/flow_base.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"math/big"
 	"time"
 
 	agglayertypes "github.com/agglayer/aggkit/agglayer/types"
@@ -555,22 +554,6 @@ type gerStatusCache struct {
 	errors     map[common.Hash]error
 }
 
-type invalidClaimAssessmentReason string
-
-const (
-	invalidClaimAssessmentReasonRecoverable invalidClaimAssessmentReason = "recoverable"
-	invalidClaimAssessmentReasonNoUnclaim   invalidClaimAssessmentReason = "no_unclaim"
-)
-
-type invalidClaimAssessment struct {
-	reason            invalidClaimAssessmentReason
-	cutBlock          uint64
-	cutClaim          *bridgesync.Claim
-	culpritClaim      *bridgesync.Claim
-	culpritUnclaim    uint64
-	hasCulpritUnclaim bool
-}
-
 // newGERStatusCache creates a new GER status cache
 func newGERStatusCache() *gerStatusCache {
 	return &gerStatusCache{
@@ -701,218 +684,4 @@ func (f *baseFlow) adjustCertificateIfNonFinalizedClaims(
 	}
 
 	return certParams, nil
-}
-
-// validateUnclaimsForUnfinalizedGERs validates unclaims for unfinalized GERs that don't exist on L1
-// in a single pass. This function combines two checks:
-//  1. Checks if any claim with unfinalized GER that doesn't exist on L1 has an unclaim that appears
-//     after a later unfinalized claim. If so, returns the block number to cut at.
-//  2. Validates that any previous claims with unfinalized GERs that don't exist on L1 have their
-//     unclaims before the current claim's block. If a claim without an unclaim is found, returns
-//     the block number to cut at.
-//
-// Returns the earliest cut block found (or 0 if no cut is needed) and an error if validation fails.
-func (f *baseFlow) validateUnclaimsForUnfinalizedGERs(
-	certParams *types.CertificateBuildParams,
-	cache *gerStatusCache) (*invalidClaimAssessment, error) {
-	// Build a map of unclaims by GlobalIndex for quick lookup
-	unclaimMap := make(map[string]uint64)
-	for _, unclaim := range certParams.Unclaims {
-		key := unclaim.GlobalIndex.String()
-		// Keep the earliest unclaim if there are multiple
-		if existing, ok := unclaimMap[key]; !ok || unclaim.BlockNumber < existing {
-			unclaimMap[key] = unclaim.BlockNumber
-		}
-	}
-
-	var recoverableClaim *invalidClaimAssessment
-
-	// Single pass through all claims to perform both checks
-	for i, claim := range certParams.Claims {
-		// Check if this claim's GER is finalized
-		isGERFinalized, err := f.getGERFinalizedStatus(cache, claim.GlobalExitRoot, certParams.L1InfoTreeLeafCount)
-		if err != nil {
-			return nil, fmt.Errorf("error checking if claim's GER %s is finalized: %w",
-				claim.GlobalExitRoot.String(), err)
-		}
-
-		// Skip if GER is finalized
-		if isGERFinalized {
-			continue
-		}
-
-		// Check 1: Validate that claims before currentBlockNum with unfinalized GERs that don't exist on L1
-		// have their unclaims before currentBlockNum
-		unclaimBlock, hasUnclaim := unclaimMap[claim.GlobalIndex.String()]
-		if !hasUnclaim {
-			currentClaim := claim
-			return &invalidClaimAssessment{
-				reason:       invalidClaimAssessmentReasonNoUnclaim,
-				cutBlock:     claim.BlockNum,
-				cutClaim:     &currentClaim,
-				culpritClaim: &currentClaim,
-			}, nil
-		}
-
-		// Check 2: Ensure we can include this claim's unclaim without being forced to include
-		// a later unfinalized claim that doesn't exist on L1 and remains active in the same
-		// certificate range.
-		//
-		// If the later claim also has an unclaim somewhere within the candidate certificate,
-		// getImportedBridgeExits will cancel that claim out regardless of the relative order of
-		// the two unclaims inside the range. What matters here is whether the later claim is
-		// still active by the end of the certificate, not whether its unclaim happens before the
-		// current claim's unclaim.
-		for j := i + 1; j < len(certParams.Claims); j++ {
-			laterClaim := certParams.Claims[j]
-			if laterClaim.BlockNum > unclaimBlock {
-				// Later claim is after the unclaim, so we can include the unclaim without including the later claim
-				continue
-			}
-
-			// Check if the later claim is unfinalized
-			isLaterGERFinalized, err := f.getGERFinalizedStatus(cache, laterClaim.GlobalExitRoot, certParams.L1InfoTreeLeafCount)
-			if err != nil {
-				return nil, fmt.Errorf("error checking if later claim's GER %s is finalized: %w",
-					laterClaim.GlobalExitRoot.String(), err)
-			}
-			if isLaterGERFinalized {
-				continue
-			}
-
-			// Later unfinalized claim that doesn't exist on L1.
-			// If it also has an unclaim within the current certificate range, both claims will be
-			// canceled out when imported bridge exits are built.
-			if _, hasLaterUnclaim := unclaimMap[laterClaim.GlobalIndex.String()]; hasLaterUnclaim {
-				continue
-			}
-			// Later claim doesn't have an unclaim in the current certificate range.
-			// We need to cut before the current claim
-			cutClaim := claim
-			blockingClaim := laterClaim
-			return &invalidClaimAssessment{
-				reason:       invalidClaimAssessmentReasonNoUnclaim,
-				cutBlock:     claim.BlockNum,
-				cutClaim:     &cutClaim,
-				culpritClaim: &blockingClaim,
-			}, nil
-		}
-
-		if recoverableClaim == nil {
-			currentClaim := claim
-			recoverableClaim = &invalidClaimAssessment{
-				reason:            invalidClaimAssessmentReasonRecoverable,
-				culpritClaim:      &currentClaim,
-				culpritUnclaim:    unclaimBlock,
-				hasCulpritUnclaim: true,
-			}
-		}
-	}
-
-	return recoverableClaim, nil
-}
-
-func (f *baseFlow) logInvalidClaimNeedsUnclaim(
-	certParams *types.CertificateBuildParams,
-	assessment *invalidClaimAssessment,
-) {
-	if assessment == nil || assessment.culpritClaim == nil {
-		return
-	}
-
-	msg := fmt.Sprintf("blocking invalid claim requires an unclaim before aggsender can proceed. %s, synced_cert_range=%d-%d",
-		formatClaimForLogs(*assessment.culpritClaim), certParams.FromBlock, certParams.ToBlock)
-	if assessment.cutClaim != nil && assessment.cutClaim.GlobalIndex != nil &&
-		assessment.culpritClaim.GlobalIndex != nil &&
-		assessment.cutClaim.GlobalIndex.Cmp(assessment.culpritClaim.GlobalIndex) != 0 {
-		msg += fmt.Sprintf(", current_cut_claim_block=%d, current_cut_claim_global_index=%s",
-			assessment.cutClaim.BlockNum, assessment.cutClaim.GlobalIndex.String())
-	}
-	f.log.Warnf("%s. No matching unclaim was found in the current DB-backed candidate certificate. An unclaim needs to happen for aggsender to get unstuck.", msg)
-}
-
-func (f *baseFlow) logLimiterBlockedInvalidClaim(
-	fullCert *types.CertificateBuildParams,
-	limitedCert *types.CertificateBuildParams,
-	limiterName string,
-) error {
-	if fullCert == nil || limitedCert == nil || limitedCert.ToBlock >= fullCert.ToBlock {
-		return nil
-	}
-
-	cache := newGERStatusCache()
-	unclaimMap := make(map[string]uint64)
-	for _, unclaim := range fullCert.Unclaims {
-		key := bigIntKey(unclaim.GlobalIndex)
-		if key == "" {
-			continue
-		}
-		if existing, ok := unclaimMap[key]; !ok || unclaim.BlockNumber < existing {
-			unclaimMap[key] = unclaim.BlockNumber
-		}
-	}
-
-	for _, claim := range fullCert.Claims {
-		if claim.BlockNum > limitedCert.ToBlock {
-			break
-		}
-
-		isGERFinalized, err := f.getGERFinalizedStatus(cache, claim.GlobalExitRoot, fullCert.L1InfoTreeLeafCount)
-		if err != nil {
-			return fmt.Errorf("error checking if claim's GER %s is finalized: %w", claim.GlobalExitRoot.String(), err)
-		}
-		if isGERFinalized {
-			continue
-		}
-
-		unclaimBlock, hasUnclaim := unclaimMap[bigIntKey(claim.GlobalIndex)]
-		if !hasUnclaim || unclaimBlock <= limitedCert.ToBlock {
-			continue
-		}
-
-		suggestion := fmt.Sprintf("increase %s so block %d fits in the same certificate as this claim", limiterName, unclaimBlock)
-		if limiterName == "MaxL2BlockNumber" {
-			suggestion = fmt.Sprintf("increase MaxL2BlockNumber to at least %d", unclaimBlock)
-		}
-
-		f.log.Warnf("%s prevents aggsender from including the unclaim that clears a blocking invalid claim. %s, required_unclaim_block=%d, full_cert_range=%d-%d, limited_cert_range=%d-%d. Suggested config change: %s.",
-			limiterName, formatClaimForLogs(claim), unclaimBlock, fullCert.FromBlock, fullCert.ToBlock, limitedCert.FromBlock, limitedCert.ToBlock, suggestion)
-		return nil
-	}
-
-	return nil
-}
-
-func formatClaimForLogs(claim bridgesync.Claim) string {
-	amount := "nil"
-	if claim.Amount != nil {
-		amount = claim.Amount.String()
-	}
-
-	return fmt.Sprintf("claim_block=%d, global_index=%s, token=%s, amount=%s",
-		claim.BlockNum, bigIntKey(claim.GlobalIndex), claim.OriginAddress.Hex(), amount)
-}
-
-func bigIntKey(value *big.Int) string {
-	if value == nil {
-		return "nil"
-	}
-
-	return value.String()
-}
-
-func logLimiterBlockedInvalidClaim(
-	base types.AggsenderFlowBaser,
-	fullCert *types.CertificateBuildParams,
-	limitedCert *types.CertificateBuildParams,
-	limiterName string,
-) {
-	concreteBaseFlow, ok := base.(*baseFlow)
-	if !ok {
-		return
-	}
-
-	if err := concreteBaseFlow.logLimiterBlockedInvalidClaim(fullCert, limitedCert, limiterName); err != nil {
-		concreteBaseFlow.log.Warnf("unable to assess non-finalized claim after %s reduction: %v", limiterName, err)
-	}
 }
